# Remove support for `vnode.state = ...`