… docs/vnodes.md
… render/tests/test-component.js