<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->
As previously discussed, this removes the `vnode._state` internal member added in #1746 to avoid breaking existing object components. Now that we're shooting for a semver-major update, let's trim this garbage! 😄 

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
See [this comment in #1727](https://github.com/MithrilJS/mithril.js/issues/1727#issuecomment-289254750) and the PR description in #1746.

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Ran the tests as usual. Nothing out of the blue popped up, so it looks like it should work. Note that the reassignment check isn't actually tested, but it was never documented in the first place, with the assumption in #1339 that it would be considered read-only.

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] I have updated `docs/change-log.md`