@isiahmeadows 2017-10-01
And some shiny perf numbers, in case you're concerned about the new sanity checks:

Chrome:

```
Before:
test-perf.js:61 rerender identical vnode x 4,088,255 ops/sec ±1.01% (53 runs sampled)
test-perf.js:61 rerender same tree x 88,768 ops/sec ±3.22% (44 runs sampled)
test-perf.js:61 construct large VDOM tree x 27,674 ops/sec ±2.67% (55 runs sampled)
test-perf.js:61 mutate styles/properties x 15,085 ops/sec ±2.88% (50 runs sampled)
test-perf.js:61 repeated trees (recycling) x 228 ops/sec ±3.77% (40 runs sampled)
test-perf.js:61 repeated trees (no recycling) x 199 ops/sec ±8.05% (43 runs sampled)
test-perf.js:67 Completed perf tests in 91215ms

After:
test-perf.js:61 rerender identical vnode x 4,083,779 ops/sec ±0.53% (55 runs sampled)
test-perf.js:61 rerender same tree x 89,641 ops/sec ±1.07% (48 runs sampled)
test-perf.js:61 construct large VDOM tree x 27,678 ops/sec ±1.09% (56 runs sampled)
test-perf.js:61 mutate styles/properties x 15,720 ops/sec ±1.57% (49 runs sampled)
test-perf.js:61 repeated trees (recycling) x 230 ops/sec ±4.20% (39 runs sampled)
test-perf.js:61 repeated trees (no recycling) x 204 ops/sec ±8.44% (43 runs sampled)
test-perf.js:67 Completed perf tests in 90960ms
```

Safari:

```
Before:
[Log] rerender identical vnode x 1,130,259 ops/sec ±8.20% (31 runs sampled) (test-perf.js, line 61)
[Log] rerender same tree x 16,245 ops/sec ±4.18% (40 runs sampled) (test-perf.js, line 61)
[Log] construct large VDOM tree x 2,476 ops/sec ±4.95% (36 runs sampled) (test-perf.js, line 61)
[Log] mutate styles/properties x 11,157 ops/sec ±1.31% (54 runs sampled) (test-perf.js, line 61)
[Log] repeated trees (recycling) x 2,441 ops/sec ±1.04% (57 runs sampled) (test-perf.js, line 61)
[Log] repeated trees (no recycling) x 2,480 ops/sec ±0.55% (59 runs sampled) (test-perf.js, line 61)
[Log] Completed perf tests in 36843ms (test-perf.js, line 67)

After:
[Log] rerender identical vnode x 1,075,703 ops/sec ±8.15% (29 runs sampled) (test-perf.js, line 61)
[Log] rerender same tree x 14,633 ops/sec ±4.08% (40 runs sampled) (test-perf.js, line 61)
[Log] construct large VDOM tree x 2,304 ops/sec ±5.75% (41 runs sampled) (test-perf.js, line 61)
[Log] mutate styles/properties x 11,157 ops/sec ±1.35% (56 runs sampled) (test-perf.js, line 61)
[Log] repeated trees (recycling) x 2,162 ops/sec ±2.66% (55 runs sampled) (test-perf.js, line 61)
[Log] repeated trees (no recycling) x 2,449 ops/sec ±1.71% (56 runs sampled) (test-perf.js, line 61)
[Log] Completed perf tests in 36895ms (test-perf.js, line 67)
```

Node:

```
Before:
rerender identical vnode x 6,313,480 ops/sec ±1.34% (89 runs sampled)
rerender same tree x 99,987 ops/sec ±1.69% (80 runs sampled)
construct large VDOM tree x 24,881 ops/sec ±1.09% (84 runs sampled)
mutate styles/properties x 46,808 ops/sec ±1.88% (82 runs sampled)
repeated trees (recycling) x 3,580 ops/sec ±2.13% (76 runs sampled)
repeated trees (no recycling) x 3,631 ops/sec ±2.07% (77 runs sampled)
Completed perf tests in 37239ms

After:
rerender identical vnode x 6,324,495 ops/sec ±1.46% (89 runs sampled)
rerender same tree x 101,641 ops/sec ±1.14% (80 runs sampled)
construct large VDOM tree x 25,018 ops/sec ±0.97% (87 runs sampled)
mutate styles/properties x 47,173 ops/sec ±1.64% (80 runs sampled)
repeated trees (recycling) x 3,646 ops/sec ±2.06% (74 runs sampled)
repeated trees (no recycling) x 3,663 ops/sec ±2.18% (75 runs sampled)
Completed perf tests in 36867ms
```