@pygy 2017-10-02
I remember we discussed generating the state of non-components vnodes lazily using a getter (and thus keeping `_state` as a private backing store)... A corresponding setter could throw. I'll give it a try and look at perf numbers.