@isiahmeadows 2017-10-02
@pygy Here's the correct perf numbers for my current patch, run locally (I just copied my previous "Before" runs):

And some shiny perf numbers, in case you're concerned about the new sanity checks:

Chrome:

```
Before:
test-perf.js:61 rerender identical vnode x 4,088,255 ops/sec ±1.01% (53 runs sampled)
test-perf.js:61 rerender same tree x 88,768 ops/sec ±3.22% (44 runs sampled)
test-perf.js:61 construct large VDOM tree x 27,674 ops/sec ±2.67% (55 runs sampled)
test-perf.js:61 mutate styles/properties x 15,085 ops/sec ±2.88% (50 runs sampled)
test-perf.js:61 repeated trees (recycling) x 228 ops/sec ±3.77% (40 runs sampled)
test-perf.js:61 repeated trees (no recycling) x 199 ops/sec ±8.05% (43 runs sampled)
test-perf.js:67 Completed perf tests in 91215ms

After:
test-perf.js:61 rerender identical vnode x 3,786,353 ops/sec ±1.37% (52 runs sampled)
test-perf.js:61 rerender same tree x 88,690 ops/sec ±4.15% (51 runs sampled)
test-perf.js:61 construct large VDOM tree x 27,623 ops/sec ±2.81% (49 runs sampled)
test-perf.js:61 mutate styles/properties x 14,817 ops/sec ±3.16% (46 runs sampled)
test-perf.js:61 repeated trees (recycling) x 216 ops/sec ±7.26% (38 runs sampled)
test-perf.js:61 repeated trees (no recycling) x 202 ops/sec ±8.65% (45 runs sampled)
test-perf.js:67 Completed perf tests in 98331ms
```

Safari:

```
Before:
[Log] rerender identical vnode x 1,130,259 ops/sec ±8.20% (31 runs sampled) (test-perf.js, line 61)
[Log] rerender same tree x 16,245 ops/sec ±4.18% (40 runs sampled) (test-perf.js, line 61)
[Log] construct large VDOM tree x 2,476 ops/sec ±4.95% (36 runs sampled) (test-perf.js, line 61)
[Log] mutate styles/properties x 11,157 ops/sec ±1.31% (54 runs sampled) (test-perf.js, line 61)
[Log] repeated trees (recycling) x 2,441 ops/sec ±1.04% (57 runs sampled) (test-perf.js, line 61)
[Log] repeated trees (no recycling) x 2,480 ops/sec ±0.55% (59 runs sampled) (test-perf.js, line 61)
[Log] Completed perf tests in 36843ms (test-perf.js, line 67)

After:
[Log] rerender identical vnode x 1,025,451 ops/sec ±9.23% (29 runs sampled) (test-perf.js, line 61)
[Log] rerender same tree x 15,795 ops/sec ±5.91% (38 runs sampled) (test-perf.js, line 61)
[Log] construct large VDOM tree x 3,270 ops/sec ±11.42% (21 runs sampled) (test-perf.js, line 61)
[Log] mutate styles/properties x 11,745 ops/sec ±1.54% (57 runs sampled) (test-perf.js, line 61)
[Log] repeated trees (recycling) x 2,232 ops/sec ±1.11% (59 runs sampled) (test-perf.js, line 61)
[Log] repeated trees (no recycling) x 2,194 ops/sec ±0.90% (60 runs sampled) (test-perf.js, line 61)
[Log] Completed perf tests in 37295ms (test-perf.js, line 67)
```

Node:

```
Before:
rerender identical vnode x 6,313,480 ops/sec ±1.34% (89 runs sampled)
rerender same tree x 99,987 ops/sec ±1.69% (80 runs sampled)
construct large VDOM tree x 24,881 ops/sec ±1.09% (84 runs sampled)
mutate styles/properties x 46,808 ops/sec ±1.88% (82 runs sampled)
repeated trees (recycling) x 3,580 ops/sec ±2.13% (76 runs sampled)
repeated trees (no recycling) x 3,631 ops/sec ±2.07% (77 runs sampled)
Completed perf tests in 37239ms

After:
rerender identical vnode x 6,236,248 ops/sec ±1.41% (90 runs sampled)
rerender same tree x 98,985 ops/sec ±1.08% (82 runs sampled)
construct large VDOM tree x 24,682 ops/sec ±0.82% (82 runs sampled)
mutate styles/properties x 46,168 ops/sec ±1.63% (82 runs sampled)
repeated trees (recycling) x 3,252 ops/sec ±4.20% (70 runs sampled)
repeated trees (no recycling) x 3,590 ops/sec ±1.88% (76 runs sampled)
Completed perf tests in 36648ms
```