@isiahmeadows 2017-10-02
@pygy I'll note that with yours, the perf hit *should* be much less if we decide to change it from an object literal to using a prototype. Note that this will be at the cost of code size, though.

I'd be okay with it if that were done, but I'm more of a +0 because of the almost inevitable code size jump.