@pygy 2017-10-03
Good point regarding the prototype... Changing to getters would however require a massive test update, since there are many tests that use POJO as vnodes...